Savepoint - Single Post page issue! I noticed that the single post page started to have some loading problems (which do not result in an error) among others. Probably due to the complexity that the code has reached and the "Callbackhell" that is being created in the interaction of posts, comments and replies. I had plans to refactor this code (because of the callbackshell) when I finished the page, but I decided to postpone it for now because of these loading problems. So the savepoint with everything working is the commit before this one. I'm only doing this because I had already made a lot of changes to the code until I had problems, so this is a savepoint before refactoring.
Change so far.
Changed the token to return the profile photo as well.
Created the transaction in the database for comment likes.
Single like constraint applied to comment likes table
Created the transaction in the database for reply likes.
Single like constraint applied to the response likes table.
Both likes endpoints were adjusted.
Created the endpoint that returns all responses to a comment.
Created the endpoint that returns whether the comment (all those displayed on the page) is liked by the logged in user.
Created the endpoint that returns whether the response (all those displayed on the page) is liked by the logged in user.
Adjustment of all routes.
Created the interaction of liking comments with a working button.
Implemented loading feedback testing in the comments section.
Created the comment interaction on the front with textarea and button partially working (until the problem mentioned above).
In the comment showing partially working (until the problem mentioned above).